+++ src/diffusers/models/__init__.py
@@ -17,7 +17,6 @@ from ..utils import is_flax_available, is_torch_available
 
 if is_torch_available():
     from .attention import Transformer2DModel
-    from .embeddings import LearnedClassifierFreeSamplingEmbeddings
     from .unet_1d import UNet1DModel
     from .unet_2d import UNet2DModel
     from .unet_2d_condition import UNet2DConditionModel
